Struct identitytoolkit3::RelyingpartyMethodsBuilder

pub struct RelyingpartyMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on relyingparty resources. It is not used directly, but through the IdentityToolkit hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate identitytoolkit3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = IdentityToolkit::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `create_auth_uri(...)`, `delete_account(...)`, `download_account(...)`, `get_account_info(...)`, `get_oob_confirmation_code(...)`, `get_public_keys(...)`, `reset_password(...)`, `set_account_info(...)`, `upload_account(...)`, `verify_assertion(...)` and `verify_password(...)`
// to build up your call.
let rb = hub.relyingparty();

Methods

impl<'a, C, NC, A> RelyingpartyMethodsBuilder<'a, C, NC, A>

fn get_oob_confirmation_code(&self, request: &Relyingparty) -> RelyingpartyGetOobConfirmationCodeMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get a code for user action confirmation.

fn get_public_keys(&self) -> RelyingpartyGetPublicKeyMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get token signing public key.

fn create_auth_uri(&self, request: &IdentitytoolkitRelyingpartyCreateAuthUriRequest) -> RelyingpartyCreateAuthUriMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates the URI used by the IdP to authenticate the user.

fn verify_assertion(&self, request: &IdentitytoolkitRelyingpartyVerifyAssertionRequest) -> RelyingpartyVerifyAssertionMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Verifies the assertion returned by the IdP.

fn upload_account(&self, request: &IdentitytoolkitRelyingpartyUploadAccountRequest) -> RelyingpartyUploadAccountMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Batch upload existing user accounts.

fn get_account_info(&self, request: &IdentitytoolkitRelyingpartyGetAccountInfoRequest) -> RelyingpartyGetAccountInfoMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the account info.

fn reset_password(&self, request: &IdentitytoolkitRelyingpartyResetPasswordRequest) -> RelyingpartyResetPasswordMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Reset password for a user.

fn download_account(&self, request: &IdentitytoolkitRelyingpartyDownloadAccountRequest) -> RelyingpartyDownloadAccountMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Batch download user accounts.

fn set_account_info(&self, request: &IdentitytoolkitRelyingpartySetAccountInfoRequest) -> RelyingpartySetAccountInfoMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Set account info for a user.

fn delete_account(&self, request: &IdentitytoolkitRelyingpartyDeleteAccountRequest) -> RelyingpartyDeleteAccountMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete user account.

fn verify_password(&self, request: &IdentitytoolkitRelyingpartyVerifyPasswordRequest) -> RelyingpartyVerifyPasswordMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Verifies the user entered password.